Commit 4 (persiapan pasang API, revisi desain)

+++ app/src/main/java/com/mamisiaga/api/APIDataClass.kt
@@ -189,13 +189,13 @@ data class PertumbuhanData(
     val age: Int,
 
     @field:SerializedName("weight")
-    val weight: Double,
+    val weight: Int,
 
     @field:SerializedName("height")
-    val height: Double,
+    val height: Int,
 
     @field:SerializedName("head_diameter")
-    val headDiameter: Double
+    val headDiameter: Int
 )
 
 data class GetImunisasiNotDoneResponse(
